Update models and factor out queries
